refactor: updated name blog route and pagination logic
