--- apps/client/app/(pages)/blog/[page]/layout.tsx
@@ -1,7 +1,0 @@
-export default async function RootLayout({
-  children,
-}: {
-  children: React.ReactNode
-}) {
-  return <div>{children}</div>
-}
